build fallbacks

+++ game/main/build.rs
@@ -34,19 +34,21 @@ where
 
 const COPY_DIR: &'static str = "assets";
 
-fn main() {
+fn main()
+{
 	// Request the output directory
-	let out = env::var("PROFILE").unwrap();
-	let out = PathBuf::from(format!("../../target/{}/{}", out, COPY_DIR));
+	if let Ok(out) = env::var("PROFILE") {
+		let out = PathBuf::from(format!("../../target/{}/{}", out, COPY_DIR));
 
-	// If it is already in the output directory, delete it and start over
-	if out.exists() {
-		fs::remove_dir_all(&out).unwrap();
+		// If it is already in the output directory, delete it and start over
+		if out.exists() {
+			fs::remove_dir_all(&out).unwrap();
+		}
+
+		// Create the out directory
+		fs::create_dir(&out).unwrap();
+
+		// Copy the directory
+		copy_dir(COPY_DIR, &out);
 	}
-
-	// Create the out directory
-	fs::create_dir(&out).unwrap();
-
-	// Copy the directory
-	copy_dir(COPY_DIR, &out);
 }
